codemod: Refactor query functions out of database
